requirements.txt
flask==2.3.3
pylint==2.17.5
anthropic==0.49.0
python-dotenv==1.1.0
gunicorn==21.2.0
werkzeug==2.3.7
markupsafe==2.1.3
itsdangerous==2.1.2
jinja2==3.1.2
click==8.1.7
blinker==1.6.2
